requirements.txt
# ClaudeFlow Dependencies

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database (Legacy - for SQLite migration only, will be removed in future)
aiosqlite>=0.19.0

# Configuration & Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Async utilities
aiofiles>=23.2.0

# Claude API (for API key authentication method)
anthropic>=0.40.0

# PTY support for Claude CLI usage command
pywinpty>=2.0.0; sys_platform == 'win32'
pexpect>=4.8.0; sys_platform != 'win32'
